some refactor of file
