+++ reader/sql/sql.go
@@ -2,12 +2,8 @@ package sql
 
 import (
 	"database/sql"
-	"encoding/binary"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"reflect"
 	"strconv"
 	"strings"
@@ -570,217 +566,6 @@ func schemaCheck(rawSchemas []string) (schemas map[string]string, err error) {
 	return
 }
 
-func restoreMeta(meta *reader.Meta, rawSqls string, magicLagDur time.Duration) (offsets []int64, sqls []string, omitMeta bool) {
-	now := time.Now().Add(-magicLagDur)
-	sqls = updateSqls(rawSqls, now)
-	omitMeta = true
-	sqlAndOffsets, length, err := meta.ReadOffset()
-	if err != nil {
-		log.Errorf("Runner[%v] %v -meta data is corrupted err:%v, omit meta data", meta.RunnerName, meta.MetaFile(), err)
-		return
-	}
-	tmps := strings.Split(sqlAndOffsets, sqlOffsetConnector)
-	if int64(len(tmps)) != 2*length || int64(len(sqls)) != length {
-		log.Errorf("Runner[%v] %v -meta file is not invalid sql meta file %v， omit meta data", meta.RunnerName, meta.MetaFile(), sqlAndOffsets)
-		return
-	}
-	omitMeta = false
-	offsets = make([]int64, length)
-	for idx, sql := range sqls {
-		syncSQL := strings.Replace(tmps[idx], "@", " ", -1)
-		offset, err := strconv.ParseInt(tmps[idx+int(length)], 10, 64)
-		if err != nil || sql != syncSQL {
-			log.Errorf("Runner[%v] %v -meta file sql is out of date %v or parse offset err %v， omit this offset", meta.RunnerName, meta.MetaFile(), syncSQL, err)
-		}
-		offsets[idx] = offset
-	}
-	return
-}
-
-// 支持 YYYY, YY, MM, DD, hh, mm, ss, M, D, h, m, s
-func convertMagic(magic string, now time.Time) (ret string) {
-	switch magic {
-	case "YYYY":
-		return fmt.Sprintf("%d", now.Year())
-	case "YY":
-		return fmt.Sprintf("%d", now.Year())[2:]
-	case "MM":
-		m := int(now.Month())
-		return fmt.Sprintf("%02d", m)
-	case "M":
-		m := int(now.Month())
-		return fmt.Sprintf("%d", m)
-	case "D":
-		d := int(now.Day())
-		return fmt.Sprintf("%d", d)
-	case "DD":
-		d := int(now.Day())
-		return fmt.Sprintf("%02d", d)
-	case "hh":
-		h := now.Hour()
-		return fmt.Sprintf("%02d", h)
-	case "h":
-		h := now.Hour()
-		return fmt.Sprintf("%d", h)
-	case "mm":
-		m := now.Minute()
-		return fmt.Sprintf("%02d", m)
-	case "m":
-		m := now.Minute()
-		return fmt.Sprintf("%d", m)
-	case "ss":
-		s := now.Second()
-		return fmt.Sprintf("%02d", s)
-	case "s":
-		s := now.Second()
-		return fmt.Sprintf("%d", s)
-	}
-	return ""
-}
-
-// 仅支持 YYYY, YY, MM, DD, hh, mm, ss，不支持 M, D, h, m, s
-func convertMagicIndex(magic string, now time.Time) (ret string, index int) {
-	switch magic {
-	case "YYYY":
-		return fmt.Sprintf("%d", now.Year()), YEAR
-	case "YY":
-		return fmt.Sprintf("%d", now.Year())[2:], YEAR
-	case "MM":
-		m := int(now.Month())
-		return fmt.Sprintf("%02d", m), MONTH
-	case "DD":
-		d := int(now.Day())
-		return fmt.Sprintf("%02d", d), DAY
-	case "hh":
-		h := now.Hour()
-		return fmt.Sprintf("%02d", h), HOUR
-	case "mm":
-		m := now.Minute()
-		return fmt.Sprintf("%02d", m), MINUTE
-	case "ss":
-		s := now.Second()
-		return fmt.Sprintf("%02d", s), SECOND
-	}
-	return "", -1
-}
-
-// 渲染魔法变量，获取渲染结果
-func goMagic(rawSql string, now time.Time) (ret string) {
-	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
-	ret = sps[0]
-	for idx := 1; idx < len(sps); idx++ {
-		idxr := strings.Index(sps[idx], ")")
-		if idxr == -1 {
-			return rawSql
-		}
-		ret += convertMagic(sps[idx][0:idxr], now)
-		if idxr+1 < len(sps[idx]) {
-			ret += sps[idx][idxr+1:]
-		}
-	}
-
-	return ret
-}
-
-type MagicRes struct {
-	timeStart   []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间位置
-	timeEnd     []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间长度
-	remainIndex []int  // 按顺序记录非时间字符串开始结束位置，去除 *
-	ret         string // 渲染结果，包含 *
-}
-
-// 渲染魔法变量
-func goMagicIndex(rawSql string, now time.Time) (MagicRes, error) {
-	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
-	var magicRes = MagicRes{
-		timeStart:   []int{-1, -1, -1, -1, -1, -1},
-		timeEnd:     make([]int, 6),
-		remainIndex: []int{0},
-		ret:         sps[0],
-	}
-	recordIndex := len(magicRes.ret)
-
-	// 没有魔法变量的情况，例如 mytest*
-	if len(sps) < 2 {
-		magicRes.remainIndex = append(magicRes.remainIndex, removeWildcards(magicRes.ret, recordIndex))
-		return magicRes, nil
-	}
-
-	magicRes.remainIndex = append(magicRes.remainIndex, recordIndex)
-	for idx := 1; idx < len(sps); idx++ {
-		idxr := strings.Index(sps[idx], ")")
-		if idxr == -1 {
-			magicRes.ret = rawSql
-			return magicRes, nil
-		}
-		spsStr := sps[idx][0:idxr]
-		if len(spsStr) < 2 {
-			magicRes.ret = rawSql
-			return magicRes, fmt.Errorf(SupportReminder)
-		}
-		res, index := convertMagicIndex(sps[idx][0:idxr], now)
-		if index == -1 {
-			magicRes.ret = rawSql
-			return magicRes, fmt.Errorf(SupportReminder)
-		}
-
-		// 记录日期起始位置
-		magicRes.timeStart[index] = recordIndex
-		magicRes.ret += res
-		recordIndex = len(magicRes.ret)
-
-		// 记录日期长度
-		magicRes.timeEnd[index] = recordIndex
-
-		if idxr+1 < len(sps[idx]) {
-			spsRemain := sps[idx][idxr+1:]
-			magicRes.ret += spsRemain
-			if spsRemain == Wildcards {
-				recordIndex = len(magicRes.ret)
-				continue
-			}
-			magicRes.remainIndex = append(magicRes.remainIndex, recordIndex)
-			magicRes.remainIndex = append(magicRes.remainIndex, removeWildcards(spsRemain, len(magicRes.ret)))
-			recordIndex = len(magicRes.ret)
-		}
-	}
-
-	return magicRes, nil
-}
-
-// 若包含通配符，字段长度相应 - 1
-func removeWildcards(checkWildcards string, length int) int {
-	if strings.Contains(checkWildcards, Wildcards) {
-		return length - 1
-	}
-	return length
-}
-
-func checkMagic(rawSql string) (valid bool) {
-	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
-	now := time.Now()
-
-	for idx := 1; idx < len(sps); idx++ {
-		idxr := strings.Index(sps[idx], ")")
-		if idxr == -1 {
-			return true
-		}
-		spsStr := sps[idx][0:idxr]
-		if len(spsStr) < 2 {
-			log.Errorf(SupportReminder)
-			return false
-		}
-
-		_, index := convertMagicIndex(sps[idx][0:idxr], now)
-		if index == -1 {
-			log.Errorf(SupportReminder)
-			return false
-		}
-	}
-
-	return true
-}
-
 func updateSqls(rawsqls string, now time.Time) []string {
 	encodedSQLs := strings.Split(rawsqls, sqlSpliter)
 	sqls := make([]string, 0)
@@ -1182,342 +967,6 @@ func (r *Reader) execReadDB(curDB string, now time.Time, recordTablesDone TableR
 		}
 	}
 	return nil
-}
-func RestoreTimestmapOffset(doneFilePath string) (time.Time, map[string]string, error) {
-	filename := fmt.Sprintf("%v.%v", reader.DoneFileName, TimestampRecordsFile)
-	cachemapfilename := fmt.Sprintf("%v.%v", reader.DoneFileName, CacheMapFile)
-
-	filePath := filepath.Join(doneFilePath, filename)
-	data, err := ioutil.ReadFile(filePath)
-	if err != nil {
-		return time.Time{}, nil, err
-	}
-	tm, err := time.Parse(time.RFC3339Nano, string(data))
-	if err != nil {
-		return tm, nil, err
-	}
-
-	cachemapfilePath := filepath.Join(doneFilePath, cachemapfilename)
-	data, err = ioutil.ReadFile(cachemapfilePath)
-	if err != nil {
-		return tm, nil, err
-	}
-	cache := make(map[string]string)
-	err = json.Unmarshal(data, &cache)
-	if err != nil {
-		return tm, nil, err
-	}
-	return tm, cache, nil
-}
-
-func WriteCacheMap(doneFilePath string, cache map[string]string) (err error) {
-	var f *os.File
-	filename := fmt.Sprintf("%v.%v", reader.DoneFileName, CacheMapFile)
-	filePath := filepath.Join(doneFilePath, filename)
-	// write to tmp file
-	f, err = os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, DefaultFilePerm)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	data, err := json.Marshal(cache)
-	if err != nil {
-		return err
-	}
-	_, err = f.Write(data)
-	if err != nil {
-		return err
-	}
-
-	return f.Sync()
-}
-
-func WriteTimestmapOffset(doneFilePath, content string) (err error) {
-	var f *os.File
-	filename := fmt.Sprintf("%v.%v", reader.DoneFileName, TimestampRecordsFile)
-	filePath := filepath.Join(doneFilePath, filename)
-	// write to tmp file
-	f, err = os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, DefaultFilePerm)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = f.Write([]byte(content))
-	if err != nil {
-		return err
-	}
-
-	return f.Sync()
-}
-
-// WriteRecordsFile 将当前文件写入donefiel中
-func WriteRecordsFile(doneFilePath, content string) (err error) {
-	var f *os.File
-	filename := fmt.Sprintf("%v.%v", reader.DoneFileName, DefaultDoneRecordsFile)
-	filePath := filepath.Join(doneFilePath, filename)
-	// write to tmp file
-	f, err = os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, DefaultFilePerm)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = f.Write([]byte(content))
-	if err != nil {
-		return err
-	}
-
-	return f.Sync()
-}
-
-func convertLong(v interface{}) (int64, error) {
-	dpv := reflect.ValueOf(v)
-	if dpv.Kind() != reflect.Ptr {
-		return 0, errors.New("scanArgs not a pointer")
-	}
-	if dpv.IsNil() {
-		return 0, errors.New("scanArgs is a nil pointer")
-	}
-	dv := reflect.Indirect(dpv)
-	switch dv.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return dv.Int(), nil
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		return int64(dv.Uint()), nil
-	case reflect.String:
-		return strconv.ParseInt(dv.String(), 10, 64)
-	case reflect.Interface:
-		idv := dv.Interface()
-		if ret, ok := idv.(int64); ok {
-			return ret, nil
-		}
-		if ret, ok := idv.(int); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(uint); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(uint64); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(string); ok {
-			return strconv.ParseInt(ret, 10, 64)
-		}
-		if ret, ok := idv.(int8); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(int16); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(int32); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(uint8); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(uint16); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.(uint32); ok {
-			return int64(ret), nil
-		}
-		if ret, ok := idv.([]byte); ok {
-			if len(ret) == 8 {
-				return int64(binary.BigEndian.Uint64(ret)), nil
-			} else {
-				return strconv.ParseInt(string(ret), 10, 64)
-			}
-		}
-		if idv == nil {
-			return 0, nil
-		}
-		log.Errorf("sql reader convertLong for type %v is not supported", reflect.TypeOf(idv))
-	}
-	return 0, fmt.Errorf("%v type can not convert to int", dv.Kind())
-}
-
-func convertFloat(v interface{}) (float64, error) {
-	dpv := reflect.ValueOf(v)
-	if dpv.Kind() != reflect.Ptr {
-		return 0, errors.New("scanArgs not a pointer")
-	}
-	if dpv.IsNil() {
-		return 0, errors.New("scanArgs is a nil pointer")
-	}
-	dv := reflect.Indirect(dpv)
-	switch dv.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return float64(dv.Int()), nil
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		return float64(dv.Uint()), nil
-	case reflect.Float32, reflect.Float64:
-		return dv.Float(), nil
-	case reflect.String:
-		return strconv.ParseFloat(dv.String(), 64)
-	case reflect.Interface:
-		idv := dv.Interface()
-		if ret, ok := idv.(float64); ok {
-			return ret, nil
-		}
-		if ret, ok := idv.(float32); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(int64); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(int); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(uint); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(uint64); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(string); ok {
-			return strconv.ParseFloat(ret, 64)
-		}
-		if ret, ok := idv.(int8); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(int16); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(int32); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(uint8); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(uint16); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.(uint32); ok {
-			return float64(ret), nil
-		}
-		if ret, ok := idv.([]byte); ok {
-			return strconv.ParseFloat(string(ret), 64)
-		}
-		if idv == nil {
-			return 0, nil
-		}
-		log.Errorf("sql reader convertFloat for type %v is not supported", reflect.TypeOf(idv))
-	}
-	return 0, fmt.Errorf("%v type can not convert to int", dv.Kind())
-}
-
-func convertDate(v interface{}) (time.Time, error) {
-	dpv := reflect.ValueOf(v)
-	if dpv.Kind() != reflect.Ptr {
-		return time.Time{}, errors.New("scanArgs not a pointer")
-	}
-	if dpv.IsNil() {
-		return time.Time{}, errors.New("scanArgs is a nil pointer")
-	}
-	dv := reflect.Indirect(dpv)
-	switch dv.Kind() {
-	case reflect.Interface:
-		idv := dv.Interface()
-		if ret, ok := idv.(time.Time); ok {
-			return ret, nil
-		}
-		if ret, ok := idv.(*time.Time); ok {
-			return *ret, nil
-		}
-		if idv == nil {
-			return time.Time{}, nil
-		}
-		log.Errorf("sql reader convertDate for type %v is not supported", reflect.TypeOf(idv))
-	}
-	return time.Time{}, fmt.Errorf("%v type can not convert to string", dv.Kind())
-}
-
-func convertString(v interface{}) (string, error) {
-	dpv := reflect.ValueOf(v)
-	if dpv.Kind() != reflect.Ptr {
-		return "", errors.New("scanArgs not a pointer")
-	}
-	if dpv.IsNil() {
-		return "", errors.New("scanArgs is a nil pointer")
-	}
-	dv := reflect.Indirect(dpv)
-	switch dv.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return strconv.Itoa(int(dv.Int())), nil
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		return strconv.Itoa(int(dv.Uint())), nil
-	case reflect.String:
-		return dv.String(), nil
-	case reflect.Interface:
-		idv := dv.Interface()
-		if ret, ok := idv.(int64); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(int); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(uint); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(uint64); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(string); ok {
-			return ret, nil
-		}
-		if ret, ok := idv.(int8); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(int16); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(int32); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(uint8); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(uint16); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.(uint32); ok {
-			return strconv.Itoa(int(ret)), nil
-		}
-		if ret, ok := idv.([]byte); ok {
-			return string(ret), nil
-		}
-		//Postgres的时间类型为time.Time
-		if ret, ok := idv.(time.Time); ok {
-			return ret.Format(time.RFC3339), nil
-		}
-		if idv == nil {
-			return "", nil
-		}
-		log.Errorf("sql reader convertString for type %v is not supported", reflect.TypeOf(idv))
-	}
-	return "", fmt.Errorf("%v type can not convert to string", dv.Kind())
-}
-
-func convertBool(v interface{}) (bool, error) {
-	dpv := reflect.ValueOf(v)
-	if dpv.Kind() != reflect.Ptr {
-		return false, errors.New("scanArgs not a pointer")
-	}
-	if dpv.IsNil() {
-		return false, errors.New("scanArgs is a nil pointer")
-	}
-	dv := reflect.Indirect(dpv)
-	switch dv.Kind() {
-	case reflect.Interface:
-		idv := dv.Interface()
-		if ret, ok := idv.(bool); ok {
-			return bool(ret), nil
-		}
-		if idv == nil {
-			return false, nil
-		}
-		log.Errorf("sql reader convertBool for type %v is not supported", reflect.TypeOf(idv))
-	}
-	return false, fmt.Errorf("%v type can not convert to Bool", dv.Kind())
 }
 
 func (r *Reader) getSQL(idx int, rawSQL string) (sql string, err error) {
@@ -2234,17 +1683,6 @@ func (r *Reader) updateStartTime(offsetKeyIndex int, scanArgs []interface{}) boo
 	}
 	return false
 }
-
-//func (r *Reader) checkTime(offsetKeyIndex int, scanArgs []interface{}) bool {
-//	timeData, ok := r.getTimeFromArgs(offsetKeyIndex, scanArgs)
-//	if ok {
-//		if timeData.After(r.startTime) {
-//			r.startTime = timeData
-//			return true
-//		}
-//	}
-//	return false
-//}
 
 func (r *Reader) trimeExistData(datas []readInfo) []readInfo {
 	if len(r.timestampKey) <= 0 || len(datas) < 1 {
